chore(pr): address Copilot review threads on dev env and tests
