+++ docker-compose.yml
@@ -84,7 +84,6 @@ services:
 volumes:
   ha-bragerone-venv:
     name: ha-bragerone-venv
-    external: true
   ha-bragerone-uv-cache:
     name: ha-bragerone-uv-cache
   ha-hass-config:
